LayoutTests/editing/input/ios/typing-with-inline-predictions-expected.txt

This test verifies that typing with inline predictions does not cause the keyboard to incorrectly autocapitalize words. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isAutoShifted is false
PASS input.value is "To whom it may concern"
PASS successfullyParsed is true

TEST COMPLETE
